requirements.txt
pydub
webrtcvad-wheels
vosk
requests
